(PC-17816) refactor(profile): move PageProfileSection from pages to components
