+++ src/features/profile/pages/NotificationSettings/NotificationSettings.tsx
@@ -7,8 +7,8 @@ import { NotificationSubscriptions, UserProfileResponse } from 'api/gen'
 import { useAuthContext } from 'features/auth/AuthContext'
 import { PushNotificationsModal } from 'features/notifications/batchNotifications/components/PushNotificationsModal'
 import { useUpdateProfileMutation } from 'features/profile/api'
+import { PageProfileSection } from 'features/profile/components/PageProfileSection/PageProfileSection'
 import { SectionWithSwitch } from 'features/profile/components/SectionWithSwitch/SectionWithSwitch'
-import { PageProfileSection } from 'features/profile/pages/PageProfileSection/PageProfileSection'
 import { useAppStateChange } from 'libs/appState'
 import { analytics } from 'libs/firebase/analytics'
 import { ButtonPrimary } from 'ui/components/buttons/ButtonPrimary'
